Move query params config to its own file
